FreeBSD libc allocator fixes.
Provide a hook so that the exported malloc symbols can be weak.  This is
mostly needed so that rtld, which statically links libc.a, can override
the symbols.

Add a bootstrap allocator so that statically linked binaries can have
their TLS space allocated before malloc is called and uses TLS.