mediumslab: limit header to min of page or slab
This had not been observed as an issue prior to
923705e514e850ee0070cef470fca9d9c47ca3da because CMakeLists.txt had, until
then, been using EQUAL, not STREQUAL, to test for oe (and to then enable
USE_SMALL_CHUNKS).  This test would fail, and so the default SLAB_SIZE was
used.  Absent this min operation, the use of a whole page on a 64KiB page
causes a crash when using the largest medium size class, as, ultimately, size
classes are not based on page sizes, and so committing a whole page to the
header leaves too little room for that class.

See also 3d3b048776fd576c725a06bd8c4432c50abca81b.