Slab metadata vs. the C++ standard
Inheritance with non-static members is not considered Standard Layout, and so
the compiler is nominally free to shuffle things around in all kinds of unusual
ways.  Instead of inheritance, rely on pointer interconvertibility with first
members of classes.

Document what's going on with static_asserts and type_traits.  The actual test
we want, std::is_pointer_interconvertible_with_class(Class::member), is not
defined until C++20 and is at least not yet in Clang.  Gate it behind the
appropriate feature test macro.