Add "wild" dimension to capptr_bounds
Introduce a notion of (nominally) allocation-bounded and exported but possibly
"wild" pointers.  These may be forged by clients and so must be validated before
they may be used as CBAllocE pointers.

capptr_dewild is a largealloc/ASM/AAL hook chain for this kind of validation.
On our sandbox test, for example, we want the "ExternalAlloc" to validate that
pointers are to within sandbox memory (which we do in the ASM).  On CHERI, we
will want to check the tag and in-bounds-ness of of any capability (and, in a
CHERI sandbox, also that it is within the sandbox region).

capptr_from_client is merely a nice bit of syntax around the CapPtr ctor.