Refactor free list to remove curr field
This shrinks the size of the metaslab as it doesn't
need to contain the current value, and uses the end value to
give this.

This has the draw back that the list can't be checked as easily
for corrupt pointers.  Once, we add termination with the bottom
bit set to the free lists, then we can move the check to
  curr -> next
in the same slab.  This currently requires a check for null.
The refactoring of initialisation of the per-thread datastructures
should enable this.

The removal allows for larger parts of the previous pointer to be stored
allowing slabs above 2^16.  We can now enable free-lists on slabs of up
to 2^32, if we resize prev to be 32bit.